+++ pyemma/coordinates/tests/test_pipeline.py
@@ -192,7 +192,7 @@ class TestPipeline(unittest.TestCase):
     def test_chunksize(self):
         reader_xtc = api.source(self.traj_files, top=self.pdb_file)
         chunksize = 1001
-        chain = [reader_xtc, api.tica(), api.cluster_mini_batch_kmeans()]
+        chain = [reader_xtc, api.tica(), api.cluster_mini_batch_kmeans(batch_size=0.3, k=3)]
         p = api.pipeline(chain, chunksize=chunksize)
         assert p.chunksize == chunksize
         for e in p._chain: